separate utils from ptable plotter
